added modifiedTimestamp filed to every graph
Since  while inside a StreamScheduler I don't know if I should modify the graph timestamp or not (I have only to do it in endless mode, and I don't want to pass this information to a Stream), the "modifiedTimestamp" is added to every graph that pass through TW by the RdfStreamScheduler stream (just before  the this.push(data) call to minize the delay). This contains value of the current timestamp retrieved used the moment() function call.